Move group fixtures to a location to be re-used
