+++ tests/functional/conftest.py
@@ -8,6 +8,7 @@ from tests.common.fixtures import es_client  # noqa: F401
 from tests.common.fixtures import init_elasticsearch  # noqa: F401
 from tests.common.fixtures.elasticsearch import ELASTICSEARCH_INDEX, ELASTICSEARCH_URL
 from tests.functional.fixtures.authentication import *  # noqa: F403,F401
+from tests.functional.fixtures.groups import *  # noqa: F403,F401
 
 TEST_SETTINGS = {
     "es.url": ELASTICSEARCH_URL,
